feat: sanitize OpenAPI tag names for Commander.js compatibility
Tags with slashes, spaces, apostrophes or other special characters
(e.g. Apify's 'Actors/Actor builds', \"Last Actor run's abort\") broke
command registration because Commander.js interprets / as path
separator and rejects spaces/quotes in command names.

Now tag names are normalized via sanitizeCommandName() to
lowercase-hyphenated form, and collisions get numeric suffixes
(-2, -3, ...). Applied in both static and dynamic command builders.

Split commander-builder.test.ts into per-concern files to stay
within the 150-line gate.

BREAKING: command group names are now always lowercase. Previously
'Pets' worked as a literal command; now it's 'pets'. This matches
standard CLI convention and the existing simplifyName output.
